feat: add support for map field types
This adds support for generating fields which are maps using the
uuid/uuids suffix and have the value type set to bytes. Key type is
anything valid in protobuf.

This commit only implements the changes needed to the core itself aswell
as the Go plugin, but not the Kotlin plugin, where it's currently not
yet available.

Closes #6.